pyproject.toml
[project]
name = "game-pitch-agent"
version = "0.4.0"
description = "ゲーム企画書生成AIエージェントシステム"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "duckduckgo-search>=7.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "python-pptx>=1.0.0",
]

[project.scripts]
game-pitch = "game_pitch_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/game_pitch_agent"]

[dependency-groups]
dev = []
